chat component completed
