+++ raaghu-components/src/rds-comp-chat/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./rds-comp-chat";
