+++ raaghu-components/src/rds-comp-chat/rds-comp-chat.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import RdsCompUserComments from './rds-comp-user-comments';
+import RdsCompUserComments from './rds-comp-chat';
 const meta: Meta<typeof RdsCompUserComments> = {
     title: "Components/Chat",
     component: RdsCompUserComments,
@@ -10,7 +10,12 @@ const meta: Meta<typeof RdsCompUserComments> = {
     tags: ['autodocs'],
     argTypes: {
         allowDelete: { control: "boolean" },
-      //  width: { control: "select", options: ["small", "medium", "large"] },
+        isEmojiPicker : {control :"boolean"},
+        isFilepload : {control :"boolean"},
+        dateFormat: {
+            options: ["dd/mm/yyyy", "mm/dd/yyyy", "Default-with-time"],
+            control: { type: "select" },
+        },
     },
 } satisfies Meta<typeof RdsCompUserComments>;
 
@@ -23,15 +28,15 @@ export const Default: Story = {
         currentUser: {
             firstName: "Max",
             lastName: "Kevin",
-            profilePic : "",
-            // profilePic :"https://cdn.shortpixel.ai/spai/q_lossy+ret_img+to_webp/honestproductreviews.com/wp-content/uploads/2018/08/Asset-5.png"
+            // If currentUser's profilePic is an empty string, display the first characters of firstName and lastName as a placeholder profile picture.
+            profilePic :"https://cdn.shortpixel.ai/spai/q_lossy+ret_img+to_webp/honestproductreviews.com/wp-content/uploads/2018/08/Asset-5.png"
         },
         comments: [
             {
                 firstName: "Alice",
                 lastName: "Smith",
                 profilePic: "https://png.pngtree.com/png-vector/20220807/ourmid/pngtree-women-avatar-wearing-suit-with-long-black-hair-png-image_6102781.png",
-                date: new Date().toLocaleDateString('en-US'),
+                date: new Date().toLocaleDateString('en-US'),  
                 comment: "Description is missing in the current version."
             },
             {
@@ -57,6 +62,9 @@ export const Default: Story = {
             },
         ],
         allowDelete: false,
+        isEmojiPicker : false ,
+        isFilepload : false,
+        dateFormat  : 'mm/dd/yyyy'
     },
 };
 
@@ -98,48 +106,8 @@ export const WithDelete: Story = {
             },
         ],
         allowDelete: true,
+        isEmojiPicker : false ,
+        isFilepload : false,
+        dateFormat  : 'mm/dd/yyyy'
     },
 };
-
-// export const LongComments: Story = {
-//     args: {
-//         currentUser: {
-//             firstName: "Max",
-//             lastName: "Parker",
-//             profilePic : ""
-
-//             //profilePic: "https://png.pngtree.com/png-vector/20220807/ourmid/pngtree-women-avatar-wearing-suit-with-long-black-hair-png-image_6102781.png",
-//         },
-//         comments: [
-//             {
-//                 firstName: "Alice",
-//                 lastName: "Smith",
-//                 profilePic: "https://png.pngtree.com/png-vector/20220807/ourmid/pngtree-women-avatar-wearing-suit-with-long-black-hair-png-image_6102781.png",
-//                 date: new Date(),
-//                 comment: "This is a great product!"
-//             },
-//             {
-//                 firstName: "Maria",
-//                 lastName: "Brown",
-//                 profilePic: "https://www.educationworld.co.in/wp-content/uploads/2023/05/dummy-female.jpg",
-//                 date: new Date(),
-//                 comment: "I had a wonderful experience."
-//             },
-//             {
-//                 firstName: "Ann",
-//                 lastName: "Augustine",
-//                 profilePic: "https://static.vecteezy.com/system/resources/previews/009/398/577/original/man-avatar-clipart-illustration-free-png.png",
-//                 date: new Date(),
-//                 comment: "I had a wonderful experience."
-//             },
-//             {
-//                 firstName: "David",
-//                 lastName: "Brown",
-//                 profilePic: "",
-//                 date: new Date(),
-//                 comment: "I had a wonderful experience."
-//             },
-//         ],
-//         allowDelete: true,
-//     },
-// };
