--- raaghu-components/src/rds-comp-user-comments/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./rds-comp-user-comments";
